3x-_ysI/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
celery==5.4.0
redis==5.0.4
httpx==0.27.0
pydantic-settings==2.3.1
